intrusion set root updated from class component to hook component
